refactor out location checking to hook and query usage
