+++ components/icons/TwitterIcon.tsx
@@ -1,4 +1,4 @@
-import { useUserStore } from "lib/stores/UserStore";
+import { useUserLocation } from "lib/hooks/useUserLocation";
 import { observer } from "mobx-react";
 
 const TwitterIcon = observer(() => {
