+++ components/icons/ZeitgeistIcon.tsx
@@ -1,4 +1,4 @@
-import { useUserStore } from "lib/stores/UserStore";
+import { useUserLocation } from "lib/hooks/useUserLocation";
 import { observer } from "mobx-react";
 
 const ZeitgeistIcon = observer(({ width = 35, height = 35, dark = false }) => {
