+++ components/ui/NotificationCenter.tsx
@@ -1,6 +1,6 @@
 import { AnimatePresence, motion } from "framer-motion";
 import { useNotificationStore } from "lib/stores/NotificationStore";
-import { useUserStore } from "lib/stores/UserStore";
+import { useUserLocation } from "lib/hooks/useUserLocation";
 import { NotificationType } from "lib/types";
 import { observer } from "mobx-react";
 import React, { FC } from "react";
@@ -14,7 +14,7 @@ const NotificationCard: FC<{
   type: NotificationType;
   dataTest?: string;
 }> = observer(({ close, timer, lifetime, content, type, dataTest }) => {
-  const userStore = useUserStore();
+  const userStore = useUserLocation();
 
   const getColor = (type: NotificationType) => {
     switch (type) {
