+++ pages/404.tsx
@@ -1,4 +1,4 @@
-import { useUserStore } from "lib/stores/UserStore";
+import { useUserLocation } from "lib/hooks/useUserLocation";
 import { observer } from "mobx-react";
 import Image from "next/image";
 import { useRouter } from "next/router";
@@ -6,7 +6,7 @@ import { useRouter } from "next/router";
 const NotFoundPage = observer(
   ({ backText, backLink }: { backText?: string; backLink?: string }) => {
     const router = useRouter();
-    const userStore = useUserStore();
+    const userStore = useUserLocation();
     const src = userStore.theme === "dark" ? "/dark-404.png" : "/light-404.png";
 
     const handleClick = () => {
